refactor: address PR review feedback for tool boundaries
Review feedback: move Music Assistant support into a built-in custom tool package instead of hardcoding it in the MCP server.

Review feedback: centralize provider runtime/auth handling so image analysis and generation reuse the same provider logic as the conversation agent.

Add tests covering Music Assistant package loading, dispatch through the custom tool loader, and shared provider runtime behavior.
