+++ custom_components/mcp_assist/config_flow.py
@@ -148,7 +148,6 @@ from .const import (
     TOOL_FAMILY_ASSIST_BRIDGE,
     TOOL_FAMILY_DEVICE,
     TOOL_FAMILY_EXTERNAL_CUSTOM,
-    TOOL_FAMILY_MUSIC_ASSISTANT,
     TOOL_FAMILY_MEMORY,
     TOOL_FAMILY_PROFILE_SETTINGS,
     TOOL_FAMILY_RECORDER,
@@ -303,7 +302,6 @@ STATIC_TOOL_FAMILY_ALPHABETICAL = [
     TOOL_FAMILY_EXTERNAL_CUSTOM,
     TOOL_FAMILY_DEVICE,
     TOOL_FAMILY_MEMORY,
-    TOOL_FAMILY_MUSIC_ASSISTANT,
     TOOL_FAMILY_RECORDER,
     TOOL_FAMILY_RESPONSE_SERVICE,
     TOOL_FAMILY_WEATHER_FORECAST,
@@ -314,7 +312,6 @@ PROFILE_DISABLE_FIELD_BY_FAMILY = {
     TOOL_FAMILY_EXTERNAL_CUSTOM: DISABLE_CUSTOM_TOOLS_FIELD,
     TOOL_FAMILY_DEVICE: DISABLE_DEVICE_FIELD,
     TOOL_FAMILY_MEMORY: DISABLE_MEMORY_FIELD,
-    TOOL_FAMILY_MUSIC_ASSISTANT: DISABLE_MUSIC_ASSISTANT_FIELD,
     TOOL_FAMILY_RECORDER: DISABLE_RECORDER_FIELD,
     TOOL_FAMILY_RESPONSE_SERVICE: DISABLE_RESPONSE_SERVICE_FIELD,
     TOOL_FAMILY_WEATHER_FORECAST: DISABLE_WEATHER_FORECAST_FIELD,
@@ -325,7 +322,6 @@ STATIC_TOOL_FAMILY_SHARED_LABELS = {
     TOOL_FAMILY_EXTERNAL_CUSTOM: "Custom Tools",
     TOOL_FAMILY_DEVICE: "Device Tools",
     TOOL_FAMILY_MEMORY: "Memory",
-    TOOL_FAMILY_MUSIC_ASSISTANT: "Music Assistant",
     TOOL_FAMILY_RECORDER: "Recorder History",
     TOOL_FAMILY_RESPONSE_SERVICE: "Response Service Reads",
     TOOL_FAMILY_WEATHER_FORECAST: "Weather Forecast",
@@ -336,7 +332,6 @@ STATIC_TOOL_FAMILY_PROFILE_DISABLE_LABELS = {
     TOOL_FAMILY_EXTERNAL_CUSTOM: "Disable Custom Tools",
     TOOL_FAMILY_DEVICE: "Disable Device Tools",
     TOOL_FAMILY_MEMORY: "Disable Memory",
-    TOOL_FAMILY_MUSIC_ASSISTANT: "Disable Music Assistant",
     TOOL_FAMILY_RECORDER: "Disable Recorder History",
     TOOL_FAMILY_RESPONSE_SERVICE: "Disable Response Service Reads",
     TOOL_FAMILY_WEATHER_FORECAST: "Disable Weather Forecast",
